use Apache BCEL parser and attributes (#172)
* use Apache BCEL parser and attributes

* simplified code and added runtime visible PARAMETER annotations

---------

+++ src/main/java/jd/core/process/writer/ConstantValueWriter.java
@@ -47,7 +47,7 @@ public final class ConstantValueWriter
 
     public static void write(
         Loader loader, Printer printer, ReferenceMap referenceMap,
-        ClassFile classFile, Constant cv, byte constantIntegerType)
+        ClassFile classFile, Constant cv, int constantIntegerType)
     {
         ConstantPool constants = classFile.getConstantPool();
